Use host compiler for verilator simulator builds
A recent change was made to have verilator runtime depend on the c++
compiler used for conda package builds.  This doesn't work so well
because that compiler doesn't see to pick up conda packages with its
builtin ldpath.  Also, our build environment 'litex-buildenv' already
requires a host c++ compiler.

This patch continues to use the conda compiler to build verilator, but
sets the verilator compiler to use a default 'g++' which will have to be
resolved on the host environment.  This may not work great for clang or
windows environments, but I think we can fix that when we get that
problem.